gets speed and time from accelerometer.

+++ Client/measuredialog.h
@@ -25,7 +25,7 @@ private:
     Accelerometer *accelerometer;
 
 
-    int time;
+    qreal time;
     qreal speed;
 
     signals:
